DV_UI_Images_Fixes: removed referenced lists from entities to improve performance, added methods according to deleted lists, count for games, user interface now counts steps for rating games dynamically
