+++ persistance/src/main/java/cz/muni/fi/pv254/dao/GameDao.java
@@ -63,4 +63,6 @@ public interface GameDao {
      * @return list of games
      */
     List<Game>findRecommendedByUser(User author);
+
+    Long countGames();
 }
