+++ persistance/src/main/java/cz/muni/fi/pv254/dao/RecommendationDao.java
@@ -56,4 +56,8 @@ public interface RecommendationDao {
     List<Recommendation> findPositive();
 
     List<Recommendation> findNegative();
+
+    List<Recommendation> findByAuthor(User author);
+
+    List<Recommendation> findByGame(Game game);
 }
