+++ service/src/main/java/cz/muni/fi/pv254/GameServiceImpl.java
@@ -103,4 +103,9 @@ public class GameServiceImpl implements GameService {
             throw new PersistenceException(ex.getMessage());
         }
     }
+
+    @Override
+    public Long countGames() {
+        return gameDao.countGames();
+    }
 }
