+++ service/src/main/java/cz/muni/fi/pv254/algorithms/contentBasedAlgorithm.java
@@ -1,4 +1,4 @@
-package cz.muni.fi.pv254.parsing;
+package cz.muni.fi.pv254.algorithms;
 
 import cz.muni.fi.pv254.dto.*;
 import cz.muni.fi.pv254.entity.Recommendation;
@@ -76,7 +76,7 @@ public class contentBasedAlgorithm {
         Set<GameDTO> bestScore = new HashSet<>();
         Map<String, Double> abScore = new HashMap<>();
         Map<String, GameDTO> uniqABScore = new HashMap<>();
-        for(RecommendationDTO reDTO: userDTO.getRecommendations()){
+        for(RecommendationDTO reDTO: new HashSet<>(recommendationFacade.findByAuthor(userDTO))){
                 if(reDTO.isVotedUp()){
                     VoteUserGames.add(reDTO.getGame());
                 }
@@ -168,7 +168,7 @@ public class contentBasedAlgorithm {
         Map<String, GameDTO> uniqABScore = new HashMap<>();
         Map<Long, Integer> countIntersection = new HashMap<>();
         Map<String, Integer> countTag = new HashMap<>();
-        for (RecommendationDTO reDTO : userDTO.getRecommendations()) {
+        for (RecommendationDTO reDTO : new HashSet<>(recommendationFacade.findByAuthor(userDTO))) {
             if (reDTO.isVotedUp()) {
                 VoteUserGames.add(reDTO.getGame());
             }
@@ -273,7 +273,7 @@ public class contentBasedAlgorithm {
         Map<String, GameDTO> uniqABScore = new HashMap<>();
         Map<Long, Integer> countIntersection = new HashMap<>();
         Map<String, Integer> countTag = new HashMap<>();
-        for (RecommendationDTO reDTO : userDTO.getRecommendations()) {
+        for (RecommendationDTO reDTO : new HashSet<>(recommendationFacade.findByAuthor(userDTO))) {
             if (reDTO.isVotedUp()) {
                 VoteUserGames.add(reDTO.getGame());
             }
