+++ web/src/main/java/cz/muni/fi/pv254/controllers/GameController.java
@@ -1,5 +1,6 @@
 package cz.muni.fi.pv254.controllers;
 
+import cz.muni.fi.pv254.algorithms.CollaborativeFiltering;
 import cz.muni.fi.pv254.dto.GameDTO;
 import cz.muni.fi.pv254.dto.GenreDTO;
 import cz.muni.fi.pv254.dto.RecommendationDTO;
@@ -9,7 +10,7 @@ import cz.muni.fi.pv254.facade.GameFacade;
 import cz.muni.fi.pv254.facade.RecommendationFacade;
 import cz.muni.fi.pv254.facade.UserFacade;
 import cz.muni.fi.pv254.parsing.App;
-import cz.muni.fi.pv254.parsing.contentBasedAlgorithm;
+import cz.muni.fi.pv254.algorithms.contentBasedAlgorithm;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -46,6 +47,9 @@ public class GameController {
     @Autowired
     private contentBasedAlgorithm contentBased;
 
+    @Autowired
+    private CollaborativeFiltering collaborative;
+
     @RequestMapping(value = "/download", method = RequestMethod.GET)
     public String Index(Model model,
                        HttpServletRequest req,
@@ -80,10 +84,10 @@ public class GameController {
 
         if (authUser != null){
 
-            UserDTO user = userFacade.findById(authUser.getId());
-            Set<RecommendationDTO> recommendations = user.getRecommendations();
+            List<RecommendationDTO> recommendations = recommendationFacade.findByAuthor(authUser);;
 
-            if (recommendations.size() >= 9){
+            Long stepCount = (gameFacade.countGames() / 10);
+            if (recommendations.size() >= stepCount){
                 redirectAttributes.addFlashAttribute("alert_danger", "Games already rated!");
                 return "redirect:/";
             }
@@ -103,6 +107,7 @@ public class GameController {
             games = games.subList(step * 10, (step * 10) + 9);
             model.addAttribute("games",games);
             model.addAttribute("step", step);
+            model.addAttribute("steps", stepCount);
             populateGenres(games, model);
 
             return "game/rate";
@@ -129,14 +134,9 @@ public class GameController {
                 throw new ResourceNotFoundException();
             recommend.setGame(game);
             recommend.setVotedUp(true);
-            recommend = recommendationFacade.add(recommend);
+            recommendationFacade.add(recommend);
 
-            game.getRecommendations().add(recommend);
-            user.getRecommendations().add(recommend);
-            gameFacade.update(game);
-            userFacade.update(user);
-
-            if (step == 8){
+            if (step == ((gameFacade.countGames() / 10) - 1)){
                 redirectAttributes.addFlashAttribute("alert_info", "Games Successfully Rated!");
                 return "redirect:/";
             }
@@ -153,10 +153,10 @@ public class GameController {
         if (authUser == null){
             return loginRedirect(redirectAttributes);
         }
-        UserDTO user = userFacade.findById(authUser.getId());
-        if (user.getRecommendations().size() < 9){
+        List<RecommendationDTO> recommendations = recommendationFacade.findByAuthor(authUser);
+        if (recommendations.size() < (gameFacade.countGames() / 10)){
             redirectAttributes.addFlashAttribute("alert_info", "Please finish rating games");
-            return "redirect:/game/rate/" + user.getRecommendations().size();
+            return "redirect:/game/rate/" + recommendations.size();
         }
 
         return "game/recommend";
@@ -166,13 +166,13 @@ public class GameController {
     public String CollaborativePearson(Model model,
                                        HttpServletRequest req,
                                        RedirectAttributes redirectAttributes){
-        String redirect = canGetRecommendation(req, redirectAttributes);
+
+        UserDTO authUser = (UserDTO) req.getSession().getAttribute("authUser");
+        String redirect = canGetRecommendation(req, redirectAttributes, recommendationFacade.findByAuthor(authUser));
         if (redirect != null)
             return redirect;
 
-        //Replace with call to collaborative pearson
-        List<GameDTO> games = gameFacade.findAll();
-        games.sort(Comparator.comparing(GameDTO::getSteamId));
+        List<GameDTO> games = collaborative.nearestNeighborIntersection(authUser.getId(), true);
         games = games.subList(0, 4);
         model.addAttribute("games",games);
 
@@ -187,13 +187,52 @@ public class GameController {
     public String CollaborativeDice(Model model,
                                        HttpServletRequest req,
                                        RedirectAttributes redirectAttributes){
-        String redirect = canGetRecommendation(req, redirectAttributes);
+        UserDTO authUser = (UserDTO) req.getSession().getAttribute("authUser");
+        String redirect = canGetRecommendation(req, redirectAttributes, recommendationFacade.findByAuthor(authUser));
         if (redirect != null)
             return redirect;
 
-        //replace with call to collaborative dice
-        List<GameDTO> games = gameFacade.findAll();
-        games.sort(Comparator.comparing(GameDTO::getSteamId));
+        List<GameDTO> games = collaborative.nearestNeighborIntersection(authUser.getId(), false);
+        games = games.subList(0, 4);
+        model.addAttribute("games",games);
+
+        if (games.size() < 10)
+            populatePictures(games, model);
+        populateGenres(games, model);
+
+        return "game/games";
+    }
+
+    @RequestMapping("/collaborativePearsonSubset")
+    public String CollaborativePearsonSubset(Model model,
+                                       HttpServletRequest req,
+                                       RedirectAttributes redirectAttributes){
+        UserDTO authUser = (UserDTO) req.getSession().getAttribute("authUser");
+        String redirect = canGetRecommendation(req, redirectAttributes, recommendationFacade.findByAuthor(authUser));
+        if (redirect != null)
+            return redirect;
+
+        List<GameDTO> games = collaborative.nearestNeighborSubset(authUser.getId(), true);
+        games = games.subList(0, 4);
+        model.addAttribute("games",games);
+
+        if (games.size() < 10)
+            populatePictures(games, model);
+        populateGenres(games, model);
+
+        return "game/games";
+    }
+
+    @RequestMapping("/collaborativeDiceSubset")
+    public String CollaborativeDiceSubset(Model model,
+                                    HttpServletRequest req,
+                                    RedirectAttributes redirectAttributes){
+        UserDTO authUser = (UserDTO) req.getSession().getAttribute("authUser");
+        String redirect = canGetRecommendation(req, redirectAttributes, recommendationFacade.findByAuthor(authUser));
+        if (redirect != null)
+            return redirect;
+
+        List<GameDTO> games = collaborative.nearestNeighborSubset(authUser.getId(), false);
         games = games.subList(0, 4);
         model.addAttribute("games",games);
 
@@ -208,14 +247,13 @@ public class GameController {
     public String DescriptionBased(Model model,
                                        HttpServletRequest req,
                                        RedirectAttributes redirectAttributes){
-        String redirect = canGetRecommendation(req, redirectAttributes);
+
+        UserDTO authUser = (UserDTO) req.getSession().getAttribute("authUser");
+        String redirect = canGetRecommendation(req, redirectAttributes, recommendationFacade.findByAuthor(authUser));
         if (redirect != null)
             return redirect;
 
-        UserDTO authUser = (UserDTO) req.getSession().getAttribute("authUser");
-
-        UserDTO user = userFacade.findById(authUser.getId());
-        Set<GameDTO> games = contentBased.recommendationByWord(user);
+        Set<GameDTO> games = contentBased.recommendationByWord(authUser);
         model.addAttribute("games",games);
 
         if (games.size() < 10)
@@ -229,13 +267,12 @@ public class GameController {
     public String GenreBasedFrequent(Model model,
                                        HttpServletRequest req,
                                        RedirectAttributes redirectAttributes){
-        String redirect = canGetRecommendation(req, redirectAttributes);
+        UserDTO authUser = (UserDTO) req.getSession().getAttribute("authUser");
+        String redirect = canGetRecommendation(req, redirectAttributes, recommendationFacade.findByAuthor(authUser));
         if (redirect != null)
             return redirect;
 
-        UserDTO authUser = (UserDTO) req.getSession().getAttribute("authUser");
-        UserDTO user = userFacade.findById(authUser.getId());
-        Set<GameDTO> games = contentBased.recommendationFrequentByTag(user);
+        Set<GameDTO> games = contentBased.recommendationFrequentByTag(authUser);
         model.addAttribute("games",games);
         if (games.size() < 10)
             populatePictures(games, model);
@@ -248,13 +285,12 @@ public class GameController {
     public String GenreBased(Model model,
                              HttpServletRequest req,
                              RedirectAttributes redirectAttributes){
-        String redirect = canGetRecommendation(req, redirectAttributes);
+        UserDTO authUser = (UserDTO) req.getSession().getAttribute("authUser");
+        String redirect = canGetRecommendation(req, redirectAttributes, recommendationFacade.findByAuthor(authUser));
         if (redirect != null)
             return redirect;
 
-        UserDTO authUser = (UserDTO) req.getSession().getAttribute("authUser");
-        UserDTO user = userFacade.findById(authUser.getId());
-        Set<GameDTO> games = contentBased.recommendationByTag(user);
+        Set<GameDTO> games = contentBased.recommendationByTag(authUser);
         model.addAttribute("games",games);
         if (games.size() < 10)
             populatePictures(games, model);
@@ -268,13 +304,12 @@ public class GameController {
                           HttpServletRequest req,
                           RedirectAttributes redirectAttributes){
 
-        String redirect = canGetRecommendation(req, redirectAttributes);
-        if (redirect != null)
-            return redirect;
 
         UserDTO authUser = (UserDTO) req.getSession().getAttribute("authUser");
-        UserDTO user = userFacade.findById(authUser.getId());
-        List<GameDTO> games = gameFacade.findRecommendedByUser(user);
+        String redirect = canGetRecommendation(req, redirectAttributes, recommendationFacade.findByAuthor(authUser));
+        if (redirect != null)
+            return redirect;
+        List<GameDTO> games = gameFacade.findRecommendedByUser(authUser);
         model.addAttribute("games",games);
         if (games.size() < 10)
             populatePictures(games, model);
@@ -290,15 +325,15 @@ public class GameController {
     }
 
     private String canGetRecommendation(HttpServletRequest req,
-                                        RedirectAttributes redirectAttributes){
+                                        RedirectAttributes redirectAttributes,
+                                        List<RecommendationDTO> recommendations){
         UserDTO authUser = (UserDTO) req.getSession().getAttribute("authUser");
         if (authUser == null){
             return loginRedirect(redirectAttributes);
         }
-        UserDTO user = userFacade.findById(authUser.getId());
-        if (user.getRecommendations().size() < 9){
+        if (recommendations.size() < (gameFacade.countGames() / 10)){
             redirectAttributes.addFlashAttribute("alert_info", "Please finish rating games");
-            return "redirect:/game/rate/" + user.getRecommendations().size();
+            return "redirect:/game/rate/" + recommendations.size();
         }
 
         return null;
